Make stat_convert.go and stat_util.go
